refactor: replace customModules folder and update paths
