+++ src/fs/copy.js
@@ -1,6 +1,6 @@
 import { cp } from "node:fs/promises";
-import { isExist } from "./customModules/isExist.js";
-import { getPath } from "./customModules/getPath.js";
+import { isExist } from "../customModules/isExist.js";
+import { getPath } from "../customModules/getPath.js";
 
 const copy = async () => {
   try {
