+++ src/fs/read.js
@@ -1,5 +1,5 @@
 import { readFile } from "node:fs/promises";
-import { getPath } from "./customModules/getPath.js";
+import { getPath } from "../customModules/getPath.js";
 
 const read = async () => {
     try {
